Move StatusType to zaino-common/status module
- Created zaino-common/src/status.rs with StatusType enum
- zaino-state/src/status.rs now only contains AtomicStatus
- AtomicStatus imports StatusType from zaino-common
